Moved test to FolderTest
